Yet another integer overflow discovered by cluster-fuzz
My earlier change (https://github.com/openssl/openssl/pull/25239)
seems to be incomplete. The 'safe' addition introduced there is
not commutative. It is able to prevent overflow when we do
a + b, but it fails to spot overflow when the values are swapped (b + a).

In a newly reported issue [1]  the clusterfuzz presents the code
with values as follows:
	ss->timeout = 0x7fffffffffffffff; (b)
	ss->time = 0x20; (a)
the current code can prevent overflow when we do a + b. but
fails to prevent overflow when we do b + a.

[1] https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=71557